% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_By_Code.R
\name{filter_By_Code}
\alias{filter_By_Code}
\title{Returns all data for a specified Continent}
\usage{
filter_By_Code(data, code)
}
\arguments{
\item{data}{The tibble from which the information will be filtered}

\item{code}{The two letter code for a specified continent}
}
\description{
Returns all data for a specified Continent
}
