% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_continent.R
\name{filter_by_continent}
\alias{filter_by_continent}
\title{Returns all vaccination data for a specified Continent}
\usage{
filter_by_continent(data, country)
}
\arguments{
\item{data}{A tibble containing relevant covid-19 vaccination information obtained from RamiKrispin's
coronavirus package.}

\item{country}{The continent of interest in the vaccination dataset}
}
\description{
Returns all vaccination data for a specified Continent
}
