% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_country.R
\name{filter_by_country}
\alias{filter_by_country}
\title{Returns all data for a specified Continent}
\usage{
filter_by_country(data, country)
}
\arguments{
\item{data}{The tibble from which the information will be filtered}

\item{country}{The country of interest}
}
\description{
Returns all data for a specified Continent
}
