% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByContinent.R, R/filterByCountry.R
\name{filterByCountry}
\alias{filterByCountry}
\title{Returns all data for a specified Continent}
\usage{
filterByCountry(data, country)

filterByCountry(data, country)
}
\arguments{
\item{data}{The tibble from which the information will be filtered}

\item{country}{The country of interest}
}
\description{
Returns all data for a specified Continent

Returns all data for a specified Continent
}
