% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByRegion.R
\name{filterByRegion}
\alias{filterByRegion}
\title{Returns all data for a specified Continent}
\usage{
filterByRegion(data, scale = "Continental", locality)
}
\arguments{
\item{data}{The tibble from which the information will be filtered}

\item{scale}{The scale of the area of interest. "Continental" refers to Continents by name,
"Code refers to Continents by letter code, and "National" refers to Countries by name. Default
is Continental}

\item{locality}{The Name of the area of interest}
}
\description{
Returns all data for a specified Continent
}
\examples{
Africa_Data <-

#
}
