% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Installs and loads the covid vaccine dataset from the RamiKrispin's package "coronavirus"}
\usage{
get_data()
}
\description{
Installs and loads the covid vaccine dataset from the RamiKrispin's package "coronavirus"
}
\examples{
# data <- getData()
# head(data)

}
