% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Installs and loads the covid vaccine dataset from the coronavirus package}
\usage{
getData()
}
\description{
Installs and loads the covid vaccine dataset from the coronavirus package
}
\examples{
# data <- getData()
# head(data)

}
