% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_vaccinated.R
\name{percent_vaccinated}
\alias{percent_vaccinated}
\title{Adds the percentage of vaccinated people to a tibble containing vaccination data.}
\usage{
percent_vaccinated(data)
}
\arguments{
\item{data}{A tibble containing relevant covid-19 vaccination information obtained from RamiKrispin's
coronavirus package.}
}
\description{
Adds the percentage of vaccinated people to a tibble containing vaccination data.
}
