% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation_to_location.R
\name{relation_to_location}
\alias{relation_to_location}
\title{Constructs a linear regression model between a vector and physical space}
\usage{
relation_to_location(data, var, lat, long)
}
\arguments{
\item{data}{Data set of interest}

\item{var}{Variable to regress against physical space}

\item{lat}{Latitude}

\item{long}{Longitude}
}
\description{
Constructs a linear regression model between a vector and physical space
}
