% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevise.R
\name{relevise}
\alias{relevise}
\title{Removes columns deemed unnecessary in vaccine dataset. Keeps the date, Continent,
Country, Total Doses Administered, Number of People with at Least One Dose,
Total Population, Latitude, and Longitude.}
\usage{
relevise(data)
}
\arguments{
\item{data}{A tibble containing relevant covid-19 vaccination information obtained from RamiKrispin's
coronavirus package.}
}
\description{
Removes columns deemed unnecessary in vaccine dataset. Keeps the date, Continent,
Country, Total Doses Administered, Number of People with at Least One Dose,
Total Population, Latitude, and Longitude.
}
