% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccinationInfo.R
\name{vaccinationInfo}
\alias{vaccinationInfo}
\title{Installs and loads the covid vaccine dataset from the coronavirus package}
\usage{
vaccinationInfo()
}
\description{
Installs and loads the covid vaccine dataset from the coronavirus package
}
\examples{
# data <- vaccinationInfo()
# head(data)

}
