% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vax_percentage.R
\name{vax_percentage}
\alias{vax_percentage}
\title{Finds the percentage of vaccinated people on a given day.}
\usage{
vax_percentage(data)
}
\description{
Finds the percentage of vaccinated people on a given day.
}
