% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualyze.R
\name{visualyze}
\alias{visualyze}
\title{Constructs a bar plot of vaccination information from a data set}
\usage{
visualyze(data)
}
\description{
Constructs a bar plot of vaccination information from a data set
}
