% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualyze_line.R
\name{visualyze_line}
\alias{visualyze_line}
\title{Constructs a bar plot of vaccination information from a data set}
\usage{
visualyze_line(data)
}
\description{
Constructs a bar plot of vaccination information from a data set
}
