% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByContinent.R
\name{filter_By_Continent}
\alias{filter_By_Continent}
\title{Returns all data for a specified Continent}
\usage{
filter_By_Continent(data, country)
}
\arguments{
\item{data}{The tibble from which the information will be filtered}

\item{country}{The continent of interest}
}
\description{
Returns all data for a specified Continent
}
