% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_country.R
\name{filter_by_country}
\alias{filter_by_country}
\title{Returns all vaccination data for a specified Country}
\usage{
filter_by_country(data, country)
}
\arguments{
\item{data}{A tibble containing relevant covid-19 vaccination information obtained from RamiKrispin's
coronavirus package.}

\item{country}{The country of interest in the vaccination Dataset}
}
\description{
Returns all vaccination data for a specified Country
}
