% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByCountry.R
\name{filter_By_Country}
\alias{filter_By_Country}
\title{Returns all data for a specified Continent}
\usage{
filter_By_Country(data, country)
}
\arguments{
\item{data}{The tibble from which the information will be filtered}

\item{country}{The country of interest}
}
\description{
Returns all data for a specified Continent
}
