% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_vaccinated.R
\name{percent_vaccinated}
\alias{percent_vaccinated}
\title{a function that adds the percent of vaccinated people based on people with at least one dose}
\usage{
percent_vaccinated(data)
}
\description{
a function that adds the percent of vaccinated people based on people with at least one dose
}
