% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualyze_line.R
\name{visualyze_line}
\alias{visualyze_line}
\title{Constructs a trend line of selected vaccination data over time.}
\usage{
visualyze_line(data, interest, dateBreaks = waiver(), dateLabels = waiver())
}
\arguments{
\item{data}{A tibble containing relevant covid-19 vaccination information obtained from RamiKrispin's
coronavirus package.}

\item{interest}{A vector of information to be plotted against time}

\item{dateBreaks}{Specifies the data breaks in the x-axis}

\item{dateLabels}{Specifies the data Labels in the x-axis}
}
\description{
Constructs a trend line of selected vaccination data over time.
}
